Move each class to its ounw file, apply proper formatting
