+++ features/mpmoviedetails/src/main/java/com/jpp/mpmoviedetails/MovieDetailsParam.kt
@@ -1,0 +1,20 @@
+package com.jpp.mpmoviedetails
+
+import android.os.Bundle
+
+/**
+ * The initialization parameter for the [MovieDetailsViewModel.onInit] method.
+ */
+data class MovieDetailsParam(
+    val movieId: Double,
+    val movieTitle: String,
+    val movieImageUrl: String
+) {
+    companion object {
+        fun fromArguments(arguments: Bundle?) = MovieDetailsParam(
+            NavigationMovieDetails.movieId(arguments).toDouble(),
+            NavigationMovieDetails.movieTitle(arguments),
+            NavigationMovieDetails.movieImageUrl(arguments)
+        )
+    }
+}
